requirements.txt
discord.py==2.6.2
python-dotenv==0.19.0